Split perf common into a util library and fix latency collection
The perf tools' shared code moves from common.h/cpp into src/perf/util
- signals, parse, latency, report, stall, net - as the silk-perf
library, each tool including its subset. The net-perf engines replace
the per-message latency vector with a log-linear histogram (~1.6%
resolution, exact moments): a 600s baseline stored ~1.2B samples and
spent ~100s sorting them at exit, misread as a teardown hang.
