+++ src/perf/CMakeLists.txt
@@ -1,5 +1,4 @@
-add_library(silk-perf STATIC common.cpp)
-target_link_libraries(silk-perf PUBLIC silk-fibers)
+add_subdirectory(util)
 
 add_executable(file-perf file-perf.cpp)
 target_link_libraries(file-perf PRIVATE silk-perf cxxopts::cxxopts)
