+++ src/perf/file-perf.cpp
@@ -1,5 +1,7 @@
-#include "common.h"
-
+#include <perf/util/latency.h>
+#include <perf/util/parse.h>
+#include <perf/util/report.h>
+#include <perf/util/signals.h>
 #include <silk/fibers/fiber.h>
 #include <silk/util/assert.h>
 #include <silk/util/crash-dumper.h>
