+++ src/perf/http-perf.cpp
@@ -1,6 +1,10 @@
-#include "common.h"
 #include "fiber-http.h"
 
+#include <perf/util/latency.h>
+#include <perf/util/net.h>
+#include <perf/util/parse.h>
+#include <perf/util/report.h>
+#include <perf/util/signals.h>
 #include <silk/fibers/fiber.h>
 #include <silk/fibers/future.h>
 #include <silk/fibers/mutex.h>
