+++ src/perf/util/net.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <cerrno>
+
+static inline bool isExpectedShutdown(int r)
+{
+    return r == ECONNRESET || r == ECANCELED || r == EBADF || r == EPIPE || r == EINVAL;
+}
